Release/2023-06-01

+++ ClinicManagement/src/Infrastructure/Fnunez.VeterinaryClinic.ClinicManagement.Infrastructure/Services/FileSystemDeleterService.cs
@@ -6,6 +6,9 @@ public class FileSystemDeleterService : IFileSystemDeleterService
 {
     public void Delete(string filePath)
     {
+        if (string.IsNullOrEmpty(filePath))
+            throw new ArgumentException($"{nameof(filePath)} is empty.");
+
         if (File.Exists(filePath))
             File.Delete(filePath);
     }
